// code/jobportal/src/app/model/jobDetails.model.ts
export interface JobDetails{

  id: string;
  title: string;
  company: string;
  location: string;
  posted: string;  
  salary: string;
  description: string;
  responsibilities: string[];
  requirements: string[];
  benefits: string;
  deadline: string;
  companyEmail: string;
  companyPhone: string;
  companyAddress: string;
  website: string;

}